+++ docs/setup/changelog.rst
@@ -21,7 +21,7 @@ Upcoming Changes (main branch)
 - Added ``pad`` function.
 - Added ``ndarray.buffer_size`` and ``ndarray.buffer_dtype`` functions for investigation of underlying buffer types.
 - Added bitwise functions (``bitwise_and``, ``bitwise_or``, ``bitwise_xor``, ``bitwise_not``, ``bitwise_left_shift``, ``bitwise_right_shift``).
-- Added ``diagonal`` and ``diag`` functions.
+- Added matrix ``diagonal``, ``diag`` and ``trace`` functions.
 
 **Changed**